participant controller tests other half

See merge request cse1105/2023-2024/teams/oopp-team-16!81

+++ server/src/main/java/server/api/ParticipantController.java
@@ -35,6 +35,13 @@ public class ParticipantController {
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Participant> deleteParticipant(@PathVariable("id") Long id) {
+
+        if(id < 0) {
+            return ResponseEntity.badRequest().build();
+        }
+        if(!db.existsById(id)) {
+            return ResponseEntity.notFound().build();
+        }
         db.deleteById(id);
         return ResponseEntity.noContent().build();
     }
@@ -52,6 +59,9 @@ public class ParticipantController {
     public ResponseEntity<Participant> updateParticipant
         (@RequestBody Participant participant, @PathVariable("id") long id) {
         if(participant == null) {
+            return ResponseEntity.badRequest().build();
+        }
+        if(id < 0) {
             return ResponseEntity.badRequest().build();
         }
         if (!db.existsById(id)){
